modulo de Productos culminado

+++ erp/src/main/java/com/store/erp/Models/RolDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class RolDTO {
 
-    private Short idRol;
+    private int idRol;
     private String descripcion;
-    
+
 }
